requirements.txt
astroid==1.3.8
Babel==2.0
cliff==1.14.0
cmd2==0.6.8
cookies==2.2.1
coverage==3.7.1
debtcollector==0.7.0
decorator==4.0.2
dj-static==0.0.6
Django==1.8.4
django-angular==0.7.15
django-appconf==1.0.1
django-compressor==1.5
django-debug-toolbar==1.3.2
django-environ==0.3.0
django-extensions==1.5.6
django-filter==0.11.0
django-grappelli==2.7.1
django-libsass==0.4
django-redis==4.2.0
django-zurb-foundation==5.5.0
djangorestframework==3.2.3
dodgy==0.1.7
factory-boy==2.5.2
freezegun==0.3.5
gitdb==0.6.4
GitPython==1.0.1
gunicorn==19.3.0
honcho==0.6.6
-e git+https://github.com/coleifer/huey.git@6a2e47485e2780ec190eef7d2ebe179b63a7c779#egg=huey-dev
ipdb==0.8.1
ipython==4.0.0
ipython-genutils==0.1.0
iso8601==0.1.10
jira==0.50
libsass==0.8.3
logilab-common==1.0.2
Markdown==2.6.2
mccabe==0.3.1
mock==1.3.0
monotonic==0.3
msgpack-python==0.4.6
netaddr==0.7.15
netifaces==0.10.4
oauthlib==1.0.3
ordereddict==1.1
oslo.config==2.3.0
oslo.i18n==2.5.0
oslo.serialization==1.8.0
oslo.utils==2.4.0
path.py==8.1
pbr==1.6.0
pep257==0.6.0
pep8==1.6.2
pep8-naming==0.3.3
pexpect==3.3
pickleshare==0.5
prettytable==0.7.2
-e git+https://github.com/antoviaque/prospector.git@a62c73b41eaee0c92bde51438268cecd13acbc18#egg=prospector-origin_pylint-extconfig-plugins
psycopg2==2.6.1
pyflakes==0.9.2
pylint==1.4.4
pylint-celery==0.3
pylint-common==0.2.1
pylint-django==0.6.1
pylint-plugin-utils==0.2.3
pyparsing==2.0.3
python-dateutil==2.4.2
python-keystoneclient==1.6.0
python-neutronclient==2.6.0
python-novaclient==2.26.0
pytz==2015.4
PyYAML==3.11
redis==2.10.3
requests==2.7.0
requests-oauthlib==0.5.0
requests-toolbelt==0.4.0
requirements-detector==0.4
responses==0.4.0
setoptconf==0.2.0
simplegeneric==0.8.1
simplejson==3.8.0
six==1.9.0
smmap==0.9.0
sockjs-tornado==1.0.1
sqlparse==0.1.16
static3==0.6.1
stevedore==1.7.0
-e git+https://github.com/antoviaque/swampdragon.git@008d9d93f9d274c8a6b606b2c5b1c9b9a09c6996#egg=SwampDragon-origin_django1.8-import-warning
tlslite==0.4.9
tornado==4.2.1
tornado-redis==2.4.18
traitlets==4.0.0
unicodecsv==0.13.0
virtualenv==13.1.2
Werkzeug==0.10.4
wrapt==1.10.5
